requirements.txt
werkzeug~=3.0.1
PyJWT~=2.8.0
Flask~=3.0.2
flask-cors
